Clean up empty files and move code
