+++ Sources/Mac/Classes/CarouselSpot.swift
@@ -221,6 +221,26 @@ open class CarouselSpot: NSObject, Gridable {
     lineView.frame.origin.y = titleView.frame.maxY + 5
     collectionView.frame.size.height = scrollView.frame.size.height + titleView.frame.size.height
   }
+
+  public func sizeForItem(at indexPath: IndexPath) -> CGSize {
+    var width = component.span > 0
+      ? collectionView.frame.width / CGFloat(component.span)
+      : collectionView.frame.width
+
+    if let layout = layout as? NSCollectionViewFlowLayout {
+      width -= layout.sectionInset.left - layout.sectionInset.right
+      width -= layout.minimumInteritemSpacing
+      width -= layout.minimumLineSpacing
+    }
+
+    if component.items[indexPath.item].size.width == 0.0 {
+      component.items[indexPath.item].size.width = width
+    }
+
+    return CGSize(
+      width: ceil(component.items[indexPath.item].size.width),
+      height: ceil(component.items[indexPath.item].size.height))
+  }
 }
 
 extension CarouselSpot: NSCollectionViewDataSource {
